refactor: api 타입 정의 파일 분리
