+++ components/CardList/CardList.tsx
@@ -1,23 +1,9 @@
+import { LinkList, Link } from '@/lib/apiTypes';
 import Card from '../Card/Card';
-import { LinkListType } from '../SearchedCardList/SearchedCardList';
 import styles from './CardList.module.css';
 
-type link = {
-  id: string;
-  title: string;
-  created_at: string;
-  description: string;
-  image_source: string;
-  url: string;
-};
-
-// 그냥 이렇게 하는게 맞는지?
-interface Props {
-  links: LinkListType | null;
-}
-
-export default function CardList({ links }: Props) {
-  const listItems = links?.map((link: link) => (
+export default function CardList({ links }: { links: LinkList }) {
+  const listItems = links?.map((link: Link) => (
     <li key={link.id}>
       <Card link={link} />
     </li>
