+++ components/SearchedCardList/SearchedCardList.tsx
@@ -1,8 +1,8 @@
 import CardList from '@/components/CardList/CardList';
-import useGetData from '@/lib/useGetData';
-import styles from './SearchedCardList.module.css';
-import EditFeatures from '../EditFeatures/EditFeatures';
 import EmptyCardList from '../EmptyCardList/EmptyCardList';
+import useGetData from '@/lib/useGetData';
+import { LinkList } from '@/lib/apiTypes';
+import styles from './SearchedCardList.module.css';
 
 interface SearchedProps {
   userId: string;
@@ -10,22 +10,9 @@ interface SearchedProps {
   searchValue: string;
 }
 
-type List = {
-  id: string;
-  created_at: string;
-  updated_at: string;
-  url: string;
-  title: string;
-  description: string;
-  image_source: string;
-  folder_id: string;
-};
-
-export type LinkListType = List[];
-
 export default function SearchedCardList({ userId, folderId = '', searchValue = '' }: SearchedProps) {
   const query = folderId ? `?folderId=${folderId}` : '';
-  const [loading, error, links] = useGetData<LinkListType>(`/users/${userId}/links${query}`, folderId);
+  const [loading, error, links] = useGetData<LinkList>(`/users/${userId}/links${query}`, folderId);
 
   if (loading) return;
   if (error) return;
